refactor: refactor file structure
