--- java1/src/main/java/org/example/ex01/hw_1.java
@@ -1,7 +1,0 @@
-package org.example.ex01;
-
-public class hw_1 {
-	public static void main(String[] args) {
-		System.out.println("Hello world!");
-	}
-}
