--- java1/src/main/java/org/example/ex01/hw_2.java
@@ -1,9 +1,0 @@
-package org.example.ex01;
-
-public class hw_2 {
-	public static void main(String[] args) {
-		int a = 10;
-		int b = 5;
-		System.out.println("a: " + a + " b: " + b);
-	}
-}
