--- java1/src/main/java/org/example/ex01/hw_3.java
@@ -1,8 +1,0 @@
-package org.example.ex01;
-
-public class hw_3 {
-	public static void main(String[] args) {
-		String name = "John";
-		System.out.println("Hello, " + name + "!");
-	}
-}
