--- java1/src/main/java/org/example/ex02/hw_2.java
@@ -1,13 +1,0 @@
-package org.example.ex02;
-
-public class hw_2 {
-	public static void main(String[] args) {
-		int a = 1;
-		float b = 1.0f;
-		String c = "hi";
-		boolean d = true;
-
-		System.out.println(a + " " + b + " " + c + " " + d);
-
-	}
-}
