--- java1/src/main/java/org/example/ex02/hw_3.java
@@ -1,8 +1,0 @@
-package org.example.ex02;
-
-public class hw_3 {
-	public static void main(String[] args) {
-		// 한 줄에 많은 변수 선언
-		short s1 = 1, s2 = 2, s3 = 3;
-	}
-}
