--- java1/src/main/java/org/example/ex03/hw_2.java
@@ -1,7 +1,0 @@
-package org.example.ex03;
-
-public class hw_2 {
-	public int ans(int a, int b){
-		return a + b;
-	}
-}
