--- java1/src/main/java/org/example/ex03/hw_3.java
@@ -1,7 +1,0 @@
-package org.example.ex03;
-
-public class hw_3 {
-	public Double calcArea(Double radius){
-		return Math.pow(radius, 2) * 3.14;
-	}
-}
