--- java1/src/main/java/org/example/ex04/hw_3.java
@@ -1,15 +1,0 @@
-package org.example.ex04;
-
-
-public class hw_3 {
-
-	private int num1;
-	private int num2;
-
-	public boolean result(int x, int y) {
-		this.num1 = x;
-		this.num2 = y;
-		return num1 % 3 == 0 && num2 % 4 == 0;
-	}
-
-}
