--- java1/src/main/java/org/example/ex04/hw_4.java
@@ -1,16 +1,0 @@
-package org.example.ex04;
-
-public class hw_4 {
-
-	private int x;
-	private int y;
-
-	public void result() {
-		this.x = 5;
-		this.y = 3;
-
-		System.out.println(x * y);
-		System.out.println(x / y);
-	}
-
-}
