--- java1/src/main/java/org/example/ex06/hw_2.java
@@ -1,15 +1,0 @@
-package org.example.ex06;
-
-// 2. 1부터 10까지의 짝수를 출력하는 프로그램을 for문을 이용하여 작성하세요.
-//
-//     (파일 형식 : hw_2.java)
-//
-public class hw_2 {
-
-	public void result() {
-		for (int i = 2; i <= 10; i += 2) {
-			System.out.println(i);
-		}
-	}
-
-}
