--- java1/src/main/java/org/example/ex08/hw_1.java
@@ -1,17 +1,0 @@
-package org.example.ex08;
-
-
-
-
-// 5. 정수를 문자열로 바꾸는 함수를 작성하세요.
-// (파일 형식 : hw_5.java)
-// 6. 문자열을 특정 공백을 기준으로 나누어서 String 배열로 반환하는 함수를 작성하세요.
-// (파일 형식 : hw_6.java)
-// 7. 입력받은 숫자를 2진수 문자열로 반환하는 함수를 작성하세요.
-// (파일 형식 : hw_7.java)
-public class hw_1 {
-
-	public String result(String temp) {
-		return temp.toUpperCase();
-	}
-}
