+++ java1/src/main/java/org/example/java01/ex01/hw_1.java
@@ -1,0 +1,7 @@
+package org.example.java01.ex01;
+
+public class hw_1 {
+	public static void main(String[] args) {
+		System.out.println("Hello world!");
+	}
+}
