+++ java1/src/main/java/org/example/java01/ex01/hw_2.java
@@ -1,0 +1,9 @@
+package org.example.java01.ex01;
+
+public class hw_2 {
+	public static void main(String[] args) {
+		int a = 10;
+		int b = 5;
+		System.out.println("a: " + a + " b: " + b);
+	}
+}
