+++ java1/src/main/java/org/example/java01/ex01/hw_3.java
@@ -1,0 +1,8 @@
+package org.example.java01.ex01;
+
+public class hw_3 {
+	public static void main(String[] args) {
+		String name = "John";
+		System.out.println("Hello, " + name + "!");
+	}
+}
