+++ java1/src/main/java/org/example/java01/ex03/hw_2.java
@@ -1,0 +1,7 @@
+package org.example.java01.ex03;
+
+public class hw_2 {
+	public int ans(int a, int b){
+		return a + b;
+	}
+}
