+++ java1/src/main/java/org/example/java01/ex03/hw_3.java
@@ -1,0 +1,7 @@
+package org.example.java01.ex03;
+
+public class hw_3 {
+	public Double calcArea(Double radius){
+		return Math.pow(radius, 2) * 3.14;
+	}
+}
