+++ java1/src/main/java/org/example/java01/ex03/hw_4.java
@@ -1,0 +1,9 @@
+package org.example.java01.ex03;
+
+import java.util.Arrays;
+
+public class hw_4 {
+	public static Integer sum(int[] arr) {
+		return Arrays.stream(arr).sum();
+	}
+}
