+++ java1/src/main/java/org/example/java01/ex04/hw_2.java
@@ -1,0 +1,16 @@
+package org.example.java01.ex04;
+
+
+public class hw_2 {
+
+	private final int x = 5;
+	private final int y = 3;
+
+	public void result() {
+		System.out.println(this.x + this.y);
+		System.out.println(this.x - this.y);
+	}
+
+
+
+}
