+++ java1/src/main/java/org/example/java01/ex04/hw_4.java
@@ -1,0 +1,16 @@
+package org.example.java01.ex04;
+
+public class hw_4 {
+
+	private int x;
+	private int y;
+
+	public void result() {
+		this.x = 5;
+		this.y = 3;
+
+		System.out.println(x * y);
+		System.out.println(x / y);
+	}
+
+}
