+++ java1/src/main/java/org/example/java01/ex04/hw_5.java
@@ -1,0 +1,13 @@
+package org.example.java01.ex04;
+
+public class hw_5 {
+
+	private int age;
+
+	public void result(int age) {
+		this.age = age;
+
+		if (age > 18) System.out.println("성인입니다.");
+
+	}
+}
