+++ java1/src/main/java/org/example/java01/ex06/hw_5.java
@@ -1,0 +1,15 @@
+package org.example.java01.ex06;
+
+
+public class hw_5 {
+
+	public void result() {
+		for (int i = 1; i < 7; i++) {
+			StringBuilder temp = new StringBuilder();
+			for (int j = 1; j <= i; j++) {
+				temp.append("*");
+			}
+			System.out.println(temp);
+		}
+	}
+}
