+++ java1/src/main/java/org/example/java01/ex08/hw_2.java
@@ -1,0 +1,10 @@
+package org.example.java01.ex08;
+
+// 2. 입력받은 문자열을 전부 소문자로 바꾸는 함수를 작성하세요.
+// (파일 형식 : hw_2.java)
+public class hw_2 {
+
+	public String result(String temp) {
+		return temp.toLowerCase();
+	}
+}
